+++ SwiftKit/Injection/InjectionOperators.swift
@@ -11,11 +11,7 @@ import Foundation
 infix operator <- { }
 
 public func <- <T>(inout injectable: T!, injector: Injector) {
-    var optionalInjectable: T? = injectable
-    
-    optionalInjectable <- injector
-    
-    injectable = optionalInjectable
+    injectable = injector.get(T.self)
 }
 
 public func <- <T>(inout injectable: T?, injector: Injector) {
@@ -23,23 +19,27 @@ public func <- <T>(inout injectable: T?, injector: Injector) {
 }
 
 public func <- <T>(inout injectableFactory: Factory<T>!, injector: Injector) {
-    var optionalInjectableFactory: Factory<T>? = injectableFactory
-    
-    optionalInjectableFactory <- injector
-    
-    injectableFactory = optionalInjectableFactory
+    injectableFactory = injector.factory(T.self)
 }
 
 public func <- <T>(inout injectableFactory: Factory<T>?, injector: Injector) {
     injectableFactory = injector.factory(T.self)
 }
 
+public func <- <T>(instance: Instance<T>, injector: Injector) {
+    injector.inject(instance)
+}
+
+public func <- <T>(instance: OptionalInstance<T>, injector: Injector) {
+    injector.inject(instance)
+}
+
+public func <- (injector: Injector, keyName: String) -> KeyedInjector {
+    return injector[keyName]
+}
+
 public func <- <T>(inout injectable: T!, injector: KeyedInjector) {
-    var optionalInjectable: T? = injectable
-    
-    optionalInjectable <- injector
-    
-    injectable = optionalInjectable
+    injectable = injector.get(T.self)
 }
 
 public func <- <T>(inout injectable: T?, injector: KeyedInjector) {
@@ -47,13 +47,46 @@ public func <- <T>(inout injectable: T?, injector: KeyedInjector) {
 }
 
 public func <- <T>(inout injectableFactory: Factory<T>!, injector: KeyedInjector) {
-    var optionalInjectableFactory: Factory<T>? = injectableFactory
-    
-    optionalInjectableFactory <- injector
-    
-    injectableFactory = optionalInjectableFactory
+    injectableFactory = injector.factory(T.self)
 }
 
 public func <- <T>(inout injectableFactory: Factory<T>?, injector: KeyedInjector) {
     injectableFactory = injector.factory(T.self)
 }
+
+public func <- <T>(instance: Instance<T>, injector: KeyedInjector) {
+    injector.inject(instance)
+}
+
+public func <- <T>(instance: OptionalInstance<T>, injector: KeyedInjector) {
+    injector.inject(instance)
+}
+
+public func <- <PARAMETERS>(injector: Injector, parameters: PARAMETERS) -> ParametrizedInjector<PARAMETERS> {
+    return ParametrizedInjector(parameters: parameters, injector: injector)
+}
+
+public func <- <T: Parametrizable>(inout injectable: T!, injector: ParametrizedInjector<T.Parameters>) {
+    injectable = injector.get(T.self)
+}
+
+public func <- <T: Parametrizable>(inout injectable: T?, injector: ParametrizedInjector<T.Parameters>) {
+    injectable = injector.get(T.self)
+}
+
+public func <- <T: Parametrizable>(inout injectable: Factory<T>!, injector: ParametrizedInjector<T.Parameters>) {
+    injectable = injector.factory(T.self)
+}
+
+public func <- <T: Parametrizable>(inout injectable: Factory<T>?, injector: ParametrizedInjector<T.Parameters>) {
+    injectable = injector.factory(T.self)
+}
+
+public func <- <T: Parametrizable>(instance: Instance<T>, injector: ParametrizedInjector<T.Parameters>) {
+    injector.inject(instance)
+}
+
+public func <- <T: Parametrizable>(instance: OptionalInstance<T>, injector: ParametrizedInjector<T.Parameters>) {
+    injector.inject(instance)
+}
+