+++ SwiftKit/Injection/OptionalInstance.swift
@@ -1,0 +1,15 @@
+//
+//  OptionalInstance.swift
+//  SwiftKit
+//
+//  Created by Tadeas Kriz on 13/10/15.
+//  Copyright © 2015 Tadeas Kriz. All rights reserved.
+//
+
+public class OptionalInstance<T>: Instance<T?> {
+    
+    public override init() {
+        super.init()
+    }
+    
+}