+++ SwiftKit/Injection/ParametrizedFactory.swift
@@ -1,0 +1,23 @@
+//
+//  ParametrizedFactory.swift
+//  SwiftKit
+//
+//  Created by Tadeas Kriz on 13/10/15.
+//  Copyright © 2015 Tadeas Kriz. All rights reserved.
+//
+
+public class ParametrizedFactory<T: ParametrizedInjectable>: Factory<T.Parameters -> T> {
+    
+    override init(closure: () -> T.Parameters -> T) {
+        super.init(closure: closure)
+    }
+    
+    override init(injector: Injector, closure: Injector -> T.Parameters -> T) {
+        super.init(injector: injector, closure: closure)
+    }
+    
+    public func create(parameters: T.Parameters) -> T {
+        return create()(parameters)
+    }
+    
+}